Samle mapping av person-/barndata fra tps-objekt til søkerinfo-objekt som returneres til frontend Personmapper var lett å blande med mapper til intenrt
