+++ src/main/kotlin/no/nav/familie/ef/søknad/mapper/SøkerinfoMapper.kt
@@ -1,13 +1,41 @@
 package no.nav.familie.ef.søknad.mapper
 
 import no.nav.familie.ef.søknad.api.dto.Søkerinfo
+import no.nav.familie.ef.søknad.api.dto.tps.Adresse
+import no.nav.familie.ef.søknad.api.dto.tps.Barn
+import no.nav.familie.ef.søknad.api.dto.tps.Person
+import no.nav.familie.ef.søknad.integration.dto.AdresseinfoDto
 import no.nav.familie.ef.søknad.integration.dto.PersoninfoDto
 import no.nav.familie.ef.søknad.integration.dto.RelasjonDto
 
 object SøkerinfoMapper {
 
     fun mapTilSøkerinfo(personinfoDto: PersoninfoDto, aktuelleBarn: List<RelasjonDto>): Søkerinfo {
-        return Søkerinfo(PersonMapper.mapTilPerson(personinfoDto),
-                         aktuelleBarn.map(PersonMapper::mapTilBarn))
+        return Søkerinfo(mapTilPerson(personinfoDto),
+                         aktuelleBarn.map(this::mapTilBarn))
+    }
+
+    fun mapTilBarn(relasjonDto: RelasjonDto): Barn {
+        return Barn(relasjonDto.ident,
+                    relasjonDto.forkortetNavn,
+                    relasjonDto.alder,
+                    relasjonDto.fødselsdato,
+                    relasjonDto.harSammeAdresse)
+    }
+
+
+    fun mapTilPerson(personinfoDto: PersoninfoDto): Person {
+        return Person(personinfoDto.ident,
+                      personinfoDto.navn.forkortetNavn,
+                      mapTilAdresse(personinfoDto.adresseinfo),
+                      personinfoDto.egenansatt?.isErEgenansatt ?: false,
+                      personinfoDto.sivilstand?.kode?.verdi ?: "",
+                      personinfoDto.statsborgerskap?.kode?.verdi ?: "")
+    }
+
+    fun mapTilAdresse(adresseinfoDto: AdresseinfoDto?): Adresse {
+        return Adresse(adresseinfoDto?.bostedsadresse?.adresse
+                       ?: "",
+                       adresseinfoDto?.bostedsadresse?.postnummer ?: "")
     }
 }
